refactor: downloadBlob helper
